ASTER019 fix: divided EditUser page to several modules
