--- src/components/elements/Headings/types.ts
@@ -1,3 +1,0 @@
-export type HeadingProps = {
-    headingTitle: string;
-}
